Update patch set 8
Patch Set 8:

Also as a meta note, I think actually not recording the threads is a much bigger change than necessary. If we go down this route I think we should still be keeping the CPU contexts for all threads, but not capturing stack contents for anything but the crashing thread.

Patch-set: 8