Update patch set 8
Patch Set 8:

I'd like us to ship something here as soon as possible.

My preference is that we minimise the size of the data as much as reasonably possible; it seems like omitting other threads or at least limiting/removing their stack data is a straightforward way to make a large difference there (excluding some data from the module list without making it useless is probably less straightforward).

But, it doesn't sound like the sizes we're likely to encounter if we don't are unreasonably huge, so if that's what folks would prefer to start with we can just go ahead without making it any smaller and see what happens in the field.

Either way, though, I'm concerned that if we just cap the size to some limit in a way that renders the resulting dump useless if the limit is hit, then we're impeding our ability to even figure out that we're producing dumps that are too big (as well as preventing ourselves analysing those crashes if reported to us). I think that if we can't currently truncate the dump in a way that doesn't render it un-decodable, then we shouldn't cap the size at all until we can; like I mentioned, microdumps weren't capped.

Patch-set: 8
Reviewer: Gerrit User 1109505 <1109505@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>