Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I'd like us to ship something here as soon as possible.
> 
> My preference is that we minimise the size of the data as much as reasonably possible; it seems like omitting other threads or at least limiting/removing their stack data is a straightforward way to make a large difference there (excluding some data from the module list without making it useless is probably less straightforward).
> 
> But, it doesn't sound like the sizes we're likely to encounter if we don't are unreasonably huge, so if that's what folks would prefer to start with we can just go ahead without making it any smaller and see what happens in the field.
> 
> Either way, though, I'm concerned that if we just cap the size to some limit in a way that renders the resulting dump useless if the limit is hit, then we're impeding our ability to even figure out that we're producing dumps that are too big (as well as preventing ourselves analysing those crashes if reported to us). I think that if we can't currently truncate the dump in a way that doesn't render it un-decodable, then we shouldn't cap the size at all until we can; like I mentioned, microdumps weren't capped.


Can't agree more, we should ship it ASAP, it has been a year since we first talked about it in the WebView team. 
Overall, our goals are

a.The size, it is our No.1 concern, the mindup log shall not overflow the buffer or be pushed out of buffer in a short period of time.
b. On the other hand, more information shall be kept providing the former (a) one is met. 

Currently, there are 3 options here.
a. Full size of minidump, it seems too much risk to get full minidump log in the field.
b. Only have exception thread, the size is controllable, will not explode with thread increase.
c. Keep the other threads' context and part of the stack, the size will increase with thread count.

I personally prefer the [b], because
- the size is controllable, it is safer than others and have very small chance to hit the cap.
- the other thread's info is not always useful, I had a short survey in MTV WebView team, it seemed no one check the other threads than exception one when investigating the crash bugs. I am not saying the other threads info is not useful, if the team finds it is useful later and we still have room, we can always add it back.
- it is relatively simpler than [c].

Sounds good?

Patch-set: 8
Reviewer: Gerrit User 1000825 <1000825@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>