Update patch set 1
Patch Set 1:

Right, OK.

The database setting is the thing that’s meant to handle upload consent. We treat this as a property of the database and not of the running crashpad_handler process so that we can turn it on and off and have the effect be “live” without having to wait for a handler restart (meaning Chrome restart).

Breakpad used to do it a different way, but Breakpad was wrong. Don’t use it for design guidance.

Patch-set: 1