Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Right, OK.
> > > 
> > > The database setting is the thing that’s meant to handle upload consent. We treat this as a property of the database and not of the running crashpad_handler process so that we can turn it on and off and have the effect be “live” without having to wait for a handler restart (meaning Chrome restart).
> > > 
> > > Breakpad used to do it a different way, but Breakpad was wrong. Don’t use it for design guidance.
> > 
> > OK, but I still need to make this work on ChromeOS. Where should I add the code if this isn't the correct place?
> 
> Well, how do you want crash_reporter to work? Do you want it to observe the pref “live” also? That’d be best, in my opinion, but I don’t know what other constraints you might have tugging at this.

How complex is the database? How complex is the code needed to interpret the database? Is the database both backwards and forwards compatible? 

If it's non-trivial, then I need to link to some parts of crashpad to read the DB. If I import crashpad into the ChromeOS repo to get its database-handling functions, then that's a different copy from the one in the Chrome repo. The two won't be in sync. Chrome might be ahead or behind ChromeOS. Is that a problem?

Patch-set: 1
Reviewer: Gerrit User 1300299 <1300299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0