[ios] Bring up first half of UncaughtExceptionHandler.
Unwind the stack looking for any exception handlers. If an exception
handler is encountered, test to see if it is a function known to catch-
and-rethrow as a "top-level" exception handler. Various routines in
UIKit do this, and it obscures the crashing stack, since the original
throw location is no longer present on the stack (just the re-throw) when
Crashpad captures the crash report.

The is a significant improvement over the alternative,
NSSetUncaughtExceptionHandler.

Change-Id: I33b46a2773e228d5ea72bf6e2053b0610d484751