[ios] Bring up first half of UncaughtExceptionHandler.
Unwind the stack looking for any exception handlers. If an exception
handler is encountered, test to see if it is a function known to catch-
and-rethrow as a "top-level" exception handler. Various routines in
UIKit do this, and it obscures the crashing stack, since the original
throw location is no longer present on the stack (just the re-throw)
when Crashpad captures the crash report.

The is a significant improvement over the alternative,
NSSetUncaughtExceptionHandler.

The details of what happens after a fatal exception is triggered are
unresolved in this CL.  For now, simply trigger a LOG(FATAL).

Bug: crashpad:31
Change-Id: Ieebc6476a0507c466c8219c10f790ec0a624e58c