Update patch set 4
Patch Set 4:

Hmm.  I looked at this again.

For (1), I don't understand why on arm64 the stack seems to end... but it turns out the real __cxa_throw runs into the same problem, and triggers a failed_throw/std::terminate.  So we can check to see if there's a handler, if there isn't, call our terminate (or whatever we decide)

For (2)... now I don't see the sinkholes as redacted.  I don't know what happened??

For (3)... still not sure what to do.

Patch-set: 4