[ios] Bring up first half of UncaughtExceptionHandler.
When code raises an Objective-C exception, unwind the stack looking for
any exception handlers. If an exception handler is encountered, test to
see if it is a function known to catch-and-rethrow 'sinkhole' exception
handler. Various routines in UIKit do this, and they obscure the crashing
stack, since the original throw location is no longer present on the
stack (just the re-throw) when Crashpad captures the crash report. In the
case of sinkholes, trigger an immediate exception to capture the original
stack.

The is an improvement over the alternative,
NSSetUncaughtExceptionHandler, which passes along the stack frames, but
not the stack (and context) itself.

The details of what happens after a fatal exception is triggered are
unresolved in this CL.  For now, simply trigger a LOG(FATAL).

Bug: crashpad:31
Change-Id: Ieebc6476a0507c466c8219c10f790ec0a624e58c