Update patch set 3
Patch Set 3:

Patching in, iOS wants:

ERROR at //snapshot/crashpad_info_client_options.cc:18:11: Include not allowed.
#include "client/crashpad_info.h"
          ^---------------------
It is not in any dependency of
  //snapshot:snapshot
The include file is in the target(s):
  //client:client
which should somehow be reachable.
___________________
ERROR at //snapshot/ios/module_snapshot_ios.h:27:11: Include not allowed.
#include "client/crashpad_info.h"
          ^---------------------
It is not in any dependency of
  //snapshot:snapshot
The include file is in the target(s):
  //client:client
which should somehow be reachable.

Patch-set: 3