fuchsia: Update build to use new "GN" SDK
This reworks the SDK pull to get the new packaging of the SDK. It also
includes pulling down the QEMU images as part of gclient with a view to
(finally!) running tests again on the CQ, but that is not done here.

BUILD.gn rules are updated to use the SDK's versions of them.

Bug: fuchsia:7802
Change-Id: Ifcb0e90b19c7eddc32a37422e6951f214dcdfd36