fuchsia: Update build to use new "GN" SDK
This reworks the SDK pull to get the new packaging of the SDK. It also
includes pulling down the QEMU images as part of gclient with a view to
(finally!) running tests again on the CQ, but that is not done here.

BUILD.gn rules are updated to use the SDK's versions of them.

This also changes the SDK to be a pinned version, specified by
build/fuchsia_sdk.version, rather than floating at head as was
previously done. I'm not sure if this is a good idea or not, but it at
least makes it less likely that the tree/tryjobs for Fuchsia will
silently break and cause annoyance for developers on other platforms.

This requires DEPS to include the changes in
https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/2165529
as well.

Bug: fuchsia:7802
Change-Id: Ifcb0e90b19c7eddc32a37422e6951f214dcdfd36