Update patch set 3
Patch Set 3:

On Monday, April 27, 2020 at 6:03:18 PM UTC+5:30, Mark Mentovai (Gerrit) wrote:

I was forced to use it because i need to close a fd before any dump collection starts. So i do that using this customhandler and then return false so that crashpad can actually install its signal handlers.

The reason I did this was, i want package crashpad along with my app and run it on all the linux distros that we support. Here the crashpad source should be compiled on the lowest kernel_version and glibc version that we support so i have compiled it in centos7.6. But the libraries that are built in centos7.6 are not working in ubuntu  18.04 with the error i pointed in the bug. The general purpose of crashpad should be that it should be independent of the distributions of linux right?

Patch-set: 3
Tag: mailMessageId=<33e35314-7a75-4daf-b373-3d37f2def58f@chromium.org>