Update patch set 2
Patch Set 2:

We have an aspirational bug somewhere to check that generated .ninja files are identical in different out dirs. It's only aspirational in that we need to write the automated check, the part where the generated ninja files are identical worked last time we checked (which was a while ago). This would break that, I think?

Patch-set: 2
Reviewer: Gerrit User 1001629 <1001629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>