Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> We have an aspirational bug somewhere to check that generated .ninja files are identical in different out dirs. It's only aspirational in that we need to write the automated check, the part where the generated ninja files are identical worked last time we checked (which was a while ago). This would break that, I think?

There’s no way for the path to be completely stable in every possible case.

The two possibilities here are that sysroot is system-absolute or that it’s source-root-absolute.

The two options here are to use rebase_path(sysroot, "") or rebase_path(sysroot, root_build_dir).

rebase_path(sysroot, "") is a no-op and is stable for a system-absolute sysroot, but it’s not stable for a source-root-absolute sysroot, because it’ll incorporate the system-absolute path of the source root, which may vary.

rebase_path(sysroot, root_build_dir) is stable for a source-root-absolute sysroot, but it’s not stable for a system-absolute sysroot, because it’ll have a variable number of dots in it depending on where the sysroot and source root are in the filesystem, relative to one another.

So yeah, that bug definitely must be aspirational. GN doesn’t have a way to express what I want to express, as far as I can tell.

To catch others up: What I really want is “make this source-relative path into something relative to root_build_dir, but leave a path that’s already absolute alone!” rebase_path doesn’t seem to have that mode, and GN doesn’t appear to have any sort of function to let me conditionally operate differently on sysroot depending on whether it’s system-absolute or source-root-absolute. While Nico and I were discussing this, I found https://gn.googlesource.com/gn/+/743970903a39f4122a6476f9e05f59b9af03fdf6/docs/reference.md#buildfile-functions-rebase_path_rebase-a-file-or-directory-to-another-location-return-value, which says “System-absolute paths will be unchanged,” which is absolutely untrue. Who’s right, the documentation or the implementation? Also, we watched https://www.youtube.com/watch?v=bB7fgX6H9_M.

Patch-set: 2