Update patch set 4
Patch Set 4:

(1 comment)

Directly testing that we find the SONAME is awkward because I don't think we have any system tools that can read the vdso's SONAME. We use dl_iterate_phdr to verify that we find all the modules the system says are loaded, and that function gives us dlpi_name, but that's the filesystem name which doesn't necessarily match the SONAME. We can write our own test code to read the dynamic array but it would also need to be smart about when to account for load bias.

This code is exercised by tests, but the only symptom was some error logs. In production, we fall back to using the mapping name if we couldn't read the SONAME. I've added some DCHECKs which would have caught this error.

Patch-set: 4
Reviewer: Gerrit User 1172779 <1172779@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>