Update patch set 7
Patch Set 7:

(17 comments)

Cool. And sorry for the delay on the review.

macOS has “pthread introspection” hooks that receive notifications of certain thread events on-thread, including thread start. Mostly this is used as an interface for debuggers. But I don’t think the Linux world generally has anything that works exactly like that. As I recall, the hook that debuggers use on Linux to be notified of new thread creation runs on the parent thread, not the new thread. Worthless for this purpose.

Patch-set: 7