fuchsia: Remove run_tests.py functionality
This removes the rotted functionality for running tests on Fuchsia. It
had previously been broken by other platform changes. Other tools are
from old SDKs are being removed too; this approach is no longer going to
work, so just remove the old code.

The preferred way is to connect via SSH to the device, however, that
using the femu.sh from the SDK

Bug: fuchsia:54031