mac: Don’t build 32-bit ProcessReaderMac support where it’s unusable
There is no possibility to run 32-bit processes on macOS 10.15 or later.
There is never any possibility to run 32-bit processes on macOS on
arm64.

This transforms ProcessReaderMac::Is64Bit into a compile-time constant
“yes”. This is a lightweight way to get much 32-bit support code removed
from optimized compiled output, including all of process_types. In an
optimized build of crashpad_handler for arm64, this is a 3% reduction
from 569kB to 552kB (-17kB).

Change-Id: I8890a170467834b99b017f1aa3dc78f3f33cd13e