linux: Correct handling of load bias
The load bias is documented to be the difference between
the preferred and actual load address for a module, but
is declared and treated as an unsigned number, regardless
of which address is greater, by relying on overflow/underflow.

ElfImageReader and DebugRendezvous both provide ways to get
the load bias for a module and are corroborated in tests.
However, the load bias computed by DebugRendezvous does
not have access to the preferred address, so there is not
enough information to determine the signedness to use with
a VMOffset.

This patch compares the load biases modulo the numeric range
for a pointer to ignore the signedness of the value.

Change-Id: I55bc49195cfb2def06777e26388380fb9bc0f710