Add a mask MinidumpCrashpadInfo to indicate valid pointer addresses.
ARM64 supports storing pointer authentication codes in the upper bits of
a pointer. This mask can be used by LLDB to mimic ptrauth_strip and strip
the pointer authentication codes. To recover an address from pointer with
an authentication code, `AND` this mask with the pointer.

If the platform does not support pointer authentication, or the range of
valid addressees for a pointer was unaccessible, this field will be 0 and
should be ignored.

Change-Id: Ie5cef90802dd1e892d456195ab8874223eac6a1b