Fix -Wc++11-narrowing in crashpad.
timeval's first member's type is __kernel_old_time_t on at least
Windows; this is a 32-bit type as opposed to time_t, which is a 64-bit
type.  As a result, casting directly to time_t results in a truncation,
triggering the warning.

It's not possible to cast directly to __kernel_old_time_t, since that
type is not exposed here.  Instead, cast to the underlying type, long,
which should work correctly on all platforms.

This has Year 2038 problems (which is why time_t is switching to
64-bit), which I suspect would be best fixed by a larger change like
moving away from using timeval anywhere (?).  I'm not really certain
what all would be involved in trying to make this whole pipeline
Y2038-safe, so I haven't attempted to tackle this.

Bug: chromium:1216696
Change-Id: Ia335a488175b1d8e851c94b56705530bb4e421af