ios: Add API to DumpWithoutCrash to a different location.
DumpWithoutCrashAndDeferProcessingAtPath(path) allows the client to
generate an intermediate dump that can later be revoked before
processing.

AddIntermediateDump(file) allows the client to decide the intermediate
dump can be processed (along with any other intermediate dumps via
ProcessIntermediateDumps())

This is useful for Chromium hang reports, which are speculatively
generated during a hang, but are later deleted if a hang recovers.

Bug: crashpad: 31
Change-Id: Ie424b375ceae3f5c0da320e766c990ea10df2f52